rename core_version -> core_version_utility
