+++ test/test_core_version_utility.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from voicevox_engine.utility.core_version import (
+from voicevox_engine.utility.core_version_utility import (
     get_latest_core_version,
     parse_core_version,
 )
