+++ voicevox_engine/utility/core_version_utility.py
